fix upvote downvote

+++ controllers/questions_controller.js
@@ -48,10 +48,24 @@ class Question {
         })
     }
     static delete(req,res) {
-        QuestionModel.deleteOne({_id:req.params.id})
-        .then(result=>{
-            console.log("masuk",req.params.id);
-            res.status(200).json({message:"question deleted",result})
+        const decoded = jwt.verify(req.headers.token,process.env.JWT_SALT)
+        QuestionModel.findById({_id:req.params.id})
+        .then(dataQuestion=>{
+            if(dataQuestion) {
+                if(dataQuestion.userId == decoded.userId) {
+                    QuestionModel.deleteOne({_id:req.params.id})
+                    .then(result=>{
+                        res.status(200).json({message:"question deleted",result})
+                    })
+                    .catch(err=>{
+                        res.status(500).json({message:err.message})
+                    })
+                }else{
+                    res.status(500).json({message:"you dont have authorized"})
+                }    
+            }else{
+                res.status(500).json({message:"question not found"})
+            }
         })
         .catch(err=>{
             res.status(500).json({message:err.message})
@@ -59,32 +73,65 @@ class Question {
     }
 
     static update(req,res) {
-        QuestionModel.updateOne({_id:req.params.id},{$set:req.body})
-        .then(result=>{
-            res.status(200).json({message:"question updated",result})
-        })
-        .catch(err=>{
-            res.status(500).json({message:err.message})
-        })
+      QuestionModel.find({_id:req.params.id})
+      .then(dataQuestion=>{
+          if(dataQuestion) {
+              const decoded = jwt.verify(req.headers.token,process.env.JWT_SALT)
+              if(dataQuestion[0].userId == decoded.userId) {
+                QuestionModel.updateOne({_id:req.params.id},{$set:req.body})
+                .then(result=>{
+                    res.status(200).json({message:"question updated",result})
+                })
+                .catch(err=>{
+                    res.status(500).json({message:err.message})
+                })
+              }else{
+                  console.log("xx")
+                  res.status(500).json({message:"you dont have authorized"})
+              }
+          }else{
+              res.status(500).json({message:"question not found"})
+          }
+      })
+      .catch(err=>{
+          res.status(500).json({message:err.message})
+      })
     }
 
     static upvote(req,res) {
-        console.log(req.params.id,"asdasdasdasd");
-        QuestionModel.findById(req.params.id)
-        .then(dataQuestion=>{
-            dataQuestion.votes.push(req.body.userId)
-            dataQuestion.save()
-            res.status(200).json(dataQuestion)
-        })
+        try {
+          const decoded = jwt.verify(req.headers.token,process.env.JWT_SALT)
+          QuestionModel.findById(req.params.id)
+          .then(dataQuestion=>{
+              dataQuestion.votes.push(decoded.userId)
+              dataQuestion.save()
+              res.status(200).json(dataQuestion)
+          }) 
+        } catch (err) {
+            res.status(500).json({message:err.message})
+        }
     }
 
     static downvote(req,res) {
+      try {
+        const decoded = jwt.verify(req.headers.token,process.env.JWT_SALT)
         QuestionModel.findById(req.params.id)
         .then(dataQuestion=>{
-            dataQuestion.votes = req.body.currVotes
+          const votesArr= dataQuestion.votes
+          const deleteIndex = votesArr.indexOf(decoded.userId)
+          if(deleteIndex > -1) {
+            votesArr.splice(deleteIndex, 1)
+            dataQuestion.votes = votesArr
             dataQuestion.save()
             res.status(200).json(dataQuestion)
+          } else {
+            res.status(200).json({message: "userId not found"})
+          }
         })
+      } catch (err) {
+          res.status(500).json({message:err.message})
+      }
+       
     }
 
 }
